test(core): colocate deepEqual tests in utils/deep-equal.test.ts
Move the deepEqual suite out of utils.test.ts into a file colocated with its
source, matching the repo convention (AGENTS.md: "vitest tests are colocated
with source") and the existing utils/*.test.ts files.
